--- src/constants/vm-status.ts
@@ -1,11 +1,0 @@
-export const printableVMStatus = {
-  Stopped: 'Stopped',
-  Migrating: 'Migrating',
-  Provisioning: 'Provisioning',
-  Starting: 'Starting',
-  Running: 'Running',
-  Paused: 'Paused',
-  Stopping: 'Stopping',
-  Terminating: 'Terminating',
-  Unknown: 'Unknown',
-};
