+++ src/views/virtualmachines/list/VirtualMachinesList.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { TFunction } from 'i18next';
 
-import { printableVMStatus } from '@kubevirt-constants/vm-status';
 import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import {
   K8sResourceCommon,
@@ -21,6 +20,18 @@ import {
 } from '@openshift-console/dynamic-plugin-sdk';
 
 import { VMStatusConditionLabelList } from './components/VMStatusConditionLabel';
+
+export const printableVMStatus = {
+  Stopped: 'Stopped',
+  Migrating: 'Migrating',
+  Provisioning: 'Provisioning',
+  Starting: 'Starting',
+  Running: 'Running',
+  Paused: 'Paused',
+  Stopping: 'Stopping',
+  Terminating: 'Terminating',
+  Unknown: 'Unknown',
+};
 
 const columns: (t: TFunction) => TableColumn<K8sResourceCommon>[] = (t) => [
   {
@@ -128,7 +139,7 @@ const VirtualMachinesList = ({ kind }: { kind: string }) => {
 
   return (
     <>
-      <ListPageHeader title={t('Virtual Machines')}>
+      <ListPageHeader title={t('Virtual Dynamic Machines')}>
         <ListPageCreate groupVersionKind={kind}>{t('Create Virtual Machine')}</ListPageCreate>
       </ListPageHeader>
       <ListPageBody>
